pyproject.toml
#poetry export --without-hashes --without-urls | awk '{ print $1 }' FS=';' > requirements.txt

[tool.poetry]
name = "sinocare-image-agent"
version = "0.1.0"
description = ""
authors = ["shukrichiu <shukrichiu@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
oss2 = "*"
pydantic = "1.10.13"
langchain = "*"
tiktoken = "*"
openai = "*"
fastapi = "*"
python-dotenv = "*"
uuid = "*"
streamlit = "*"
streamlit-feedback = "*"
langserve = { version = "*", extras = ["all"] }
langchain-openai = "*"
langgraph = "*"
langchain-cli = "*"
langchain-community = "*"
langchain_experimental = "*"
gunicorn = "*"
dashscope = "*"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.2"
ruff = "<0.2"
mypy = "*"
black = "*"
pre-commit = "*"
importlib-metadata = "*"
streamlit = "*"
streamlit-feedback = "*"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["F401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
